requirements-airflow-runtime.txt
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
scipy==1.14.1

lightgbm==4.5.0
xgboost==2.1.1
mlflow==2.17.0
shap==0.46.0
evidently==0.4.39

pydantic==2.9.2
pydantic-settings==2.5.2

boto3==1.35.36
s3fs==2024.9.0
psycopg2-binary==2.9.9

pyyaml==6.0.2
python-dotenv==1.0.1
joblib==1.4.2
requests==2.32.3
matplotlib==3.9.2

apache-airflow==2.10.2
apache-airflow-providers-postgres==5.13.0
apache-airflow-providers-docker==3.14.0
